Extract view model and show painting cost
